+++ frontend/src/queries/useBakerQuery.ts
@@ -1,23 +1,20 @@
 import { useQuery, gql } from '@urql/vue'
 import { useComponentState } from '~/composables/useComponentState'
-// import type {
-// 	Baker,
-// 	BakerState,
-// 	ActiveBakerState,
-// 	PoolApy,
-// } from '~/types/generated'
-import type { Baker } from '~/types/generated'
+import type {
+	Baker,
+	BakerState,
+	ActiveBakerState,
+	PoolApy,
+} from '~/types/generated'
 
-export type BakerWithAPYFilter = Baker
-// TODO add this back once new rust-backend here once APY
-// & {
-// 	state: BakerState & {
-// 		pool?: ActiveBakerState['pool'] & {
-// 			apy7days: PoolApy
-// 			apy30days: PoolApy
-// 		}
-// 	}
-// }
+export type BakerWithAPYFilter = Baker & {
+	state: BakerState & {
+		pool?: ActiveBakerState['pool'] & {
+			apy7days: PoolApy
+			apy30days: PoolApy
+		}
+	}
+}
 
 type BakerResponse = {
 	bakerByBakerId: BakerWithAPYFilter
@@ -94,6 +91,16 @@ const BakerQuery = gql<BakerResponse>`
 							transactionCommission
 							bakingCommission
 						}
+						apy7days: apy(period: LAST7_DAYS) {
+							bakerApy
+							delegatorsApy
+							totalApy
+						}
+						apy30days: apy(period: LAST30_DAYS) {
+							bakerApy
+							delegatorsApy
+							totalApy
+						}
 					}
 					pendingChange {
 						... on PendingBakerReduceStake {
@@ -114,17 +121,6 @@ const BakerQuery = gql<BakerResponse>`
 		}
 	}
 `
-// TODO add this back once new rust-backend here once APY
-// apy7days: apy(period: LAST7_DAYS) {
-// 	bakerApy
-// 	delegatorsApy
-// 	totalApy
-// }
-// apy30days: apy(period: LAST30_DAYS) {
-// 	bakerApy
-// 	delegatorsApy
-// 	totalApy
-// }
 
 export const useBakerQuery = (bakerId: number) => {
 	const { data, fetching, error } = useQuery({